+++ install/finish.php
@@ -1,0 +1,21 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Install USOC</title>
+    <link href="css.css" rel="stylesheet" />
+</head>
+<body>   
+    <form action="../login.php" method="post" id="form">
+        <p id="p1">Finished!</p>
+        <div id="p2">
+            <p>Thank you for installing USOC</p>
+        </div>
+    </form>
+    <div id="p3">
+        <p>You can login with the Username Admin and your set password</p>
+        <input class="p3" type="submit" form="form" style="top:120%;" value="Go to login →" />
+    </div>
+</body>
+</html>