+++ install/step1.php
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Install USOC</title>
+    <link href="css.css" rel="stylesheet" />
+</head>
+<body>
+    <p id="p1">USOC</p>
+    <p id="p2">Install wizzard</p>
+    <button id="p3" onClick="location.href = 'step2.php'">Next step →</button>
+</body>
+</html>